Add Explore game class and ExploreJulia
The new classes implements a new iterative method to explore Julia sets.
This methods contain more detail within the center of the fractal whilst the previous method only managed to display the edges.
To display this, a new putPixel method with different paramaters was added to the ChaosCanvas which put the rate of converging as the value of a pixel.