Implement improved GUI for ChaosGame and ExploreGame
To display explore Julia sets better and faster a IntStream was implemented to compute the sets faster.
The boundaries in the while-loop was edited to work correctly. And the canvas int array was changed to a
double array. A few more methods was added to Vector2D.
Buttons to zoom and the ability to drag and move the fractal was added to ExplorePage.
A ComboBox was added to ChaosPage giving the ability to select different fractals to run as a Chaos Game.