Add open from file button in ChaosGame
Add button that when pressed opens a FileChooser that filters through and
only allows to open .txt files. The game is updated if a correctly formatted
text file is selected. In addition, the drag and move event in ExplorePage
is improved to not render through entire drag movement. Now the image only renders
when the user releases the image.